modify EntryThreshold type

+++ AElf.Kernel/TxMemPool/ITxPoolConfig.cs
@@ -20,6 +20,6 @@
         /// <summary>
         /// minimal number of txs for entering pool
         /// </summary>
-        int EntryThreshold { get; }
+        ulong EntryThreshold { get; }
     }
 }
